refactor(config): deduplicate FEEDBACK_FAILURE_MODES to schema layer
