+++ src/core/config/config-schema.ts
@@ -57,9 +57,25 @@ const httpUrl = z.string().refine((v) => v.startsWith("http://") || v.startsWith
   message: "endpoint must start with http:// or https://",
 });
 
-// ── Feedback failure modes ──────────────────────────────────────────────────
+// ── Feedback failure modes (F-3 / #384) ─────────────────────────────────────
 
-export const FEEDBACK_FAILURE_MODES = ["incorrect", "outdated", "dangerous", "incomplete", "redundant"] as const;
+/**
+ * Curated taxonomy of failure modes for negative feedback.
+ *
+ * Structured failure modes enable aggregation across feedback events so the
+ * distill pipeline can detect that "5 assets failed for the same reason" and
+ * act on it — free-text strings about the same issue are not aggregatable.
+ */
+export const FEEDBACK_FAILURE_MODES = [
+  "incorrect", // Factually wrong or logically flawed content
+  "outdated", // Correct at some point but now stale
+  "dangerous", // Could cause harm if followed (security, safety)
+  "incomplete", // Missing key steps, context, or caveats
+  "redundant", // Duplicates another asset without adding value
+] as const;
+
+/** Union of the curated failure-mode values. */
+export type FeedbackFailureMode = (typeof FEEDBACK_FAILURE_MODES)[number];
 
 // ── Connection configs (LLM / embedding) ────────────────────────────────────
